keymap_support: Update keymap support to support users  better.
Worknig with SiriusStarr, there were some problems in the initial
implementation.  We now provide a headerfile for people to use
as well as the .c, depending the user's needs.
